refactor: 코드 중복 줄이기
