+++ src/app/board/components/JoinBottomSheet/JoinBottomSheet.css.ts
@@ -19,31 +19,6 @@ export const content = style({
   gap: space['2xl'],
 });
 
-export const buttonGroup = style({
-  display: 'flex',
-  gap: space['2xs'],
-  alignItems: 'center',
-});
-
-export const lineBreak = style({
-  wordBreak: 'break-all',
-  whiteSpace: 'break-spaces',
-});
-
-export const bulletTitleBottomSpace = style({
-  marginBottom: space['sm'],
-});
-
-export const bodyBottomSpace = style({
-  marginBottom: space.lg,
-});
-
-export const checkBoxGroup = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: space.md,
-});
-
 export const footer = style({
   paddingBottom: space.xl,
 });
