+++ src/app/board/components/JoinModal/JoinModal.css.ts
@@ -15,28 +15,3 @@ export const modalContent = style({
   textAlign: 'left',
   gap: space['2xl'],
 });
-
-export const buttonGroup = style({
-  display: 'flex',
-  gap: space['2xs'],
-  alignItems: 'center',
-});
-
-export const lineBreak = style({
-  wordBreak: 'break-all',
-  whiteSpace: 'break-spaces',
-});
-
-export const bulletTitleBottomSpace = style({
-  marginBottom: space['sm'],
-});
-
-export const bodyBottomSpace = style({
-  marginBottom: space.lg,
-});
-
-export const checkBoxGroup = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: space.md,
-});
