+++ src/app/board/components/index.ts
@@ -25,3 +25,5 @@ export { default as DetailMenuButton } from './DetailMenuButton/DetailMenuButton
 export { default as BottomSection } from './BottomSection/BottomSection';
 
 export { default as JoinBottomSheet } from './JoinBottomSheet/JoinBottomSheet';
+
+export { default as JoinForm } from './JoinForm/JoinForm';
